# todo-app/backend/src/main/java/com/todo/todo/models/Todo.java
package com.todo.todo.models;

import com.fasterxml.jackson.annotation.JsonFormat;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import java.sql.Date;

@Entity
public class Todo {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Integer id;

    private String title;

    @JsonFormat(pattern = "yyyy-MM-dd", timezone = "Europe/Madrid")
    private Date created;

    private Boolean done;

    private Integer userid;


    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public Date getCreated() { return created; }

    public void setCreated(Date created) { this.created = created; }

    public Boolean getDone() { return done; }

    public void setDone(Boolean done) { this.done = done; }

    public Integer getUserId() {
        return userid;
    }

    public void setUserId(Integer userid) {
        this.userid = userid;
    }
}